QtTranslationsLoader class renamed to QtFileLoader.
