+++ src/Symfony/Component/Translation/CHANGELOG.md
@@ -4,11 +4,12 @@ CHANGELOG
 2.2.0
 -----
 
+ * QtTranslationsLoader class renamed to QtFileLoader. QtTranslationsLoader is deprecated and will be removed in 2.3.
  * [BC BREAK] uniformized the exception thrown by the load() method when an error occurs. The load() method now
    throws Symfony\Component\Translation\Exception\NotFoundResourceException when a resource cannot be found
    and Symfony\Component\Translation\Exception\InvalidResourceException when a resource is invalid.
  * changed the exception class thrown by some load() methods from \RuntimeException to \InvalidArgumentException
-   (IcuDatFileLoader, IcuResFileLoader, and QtTranslationsLoader)
+   (IcuDatFileLoader, IcuResFileLoader and QtFileLoader)
 
 2.1.0
 -----
